primer/mcp-servers/shopify-mcp/requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
ShopifyAPI
